requirements.txt
flask==3.0.3
flask-login==0.6.3
supabase==2.5.0
cloudinary==1.40.0
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==22.0.0
